+++ src/lib-cuda/reg_flow_rls.cpp
@@ -51,7 +51,7 @@ namespace {
 // int -> cudaStream_t (0 == default stream). The public ABI carries the stream
 // as an int; the cuda-impl launchers take a cudaStream_t. Mirrors
 // pushpull::_pp_stream in the cuda-impl layer.
-static inline cudaStream_t _reg_stream(int stream)
+static inline cudaStream_t _reg_stream(intptr_t stream)
 {
     return reinterpret_cast<cudaStream_t>(static_cast<std::intptr_t>(stream));
 }
@@ -303,7 +303,7 @@ void flow_matvec_rls(
           double     div       ,
           int8_t     bound     ,
           int        ndim      ,
-          int        stream
+          intptr_t   stream
 )
 {
     flow_rls_check_bending(bending, "flow_matvec_rls");
@@ -353,7 +353,7 @@ void flow_diag_rls(
           double     div       ,
           int8_t     bound     ,
           int        ndim      ,
-          int        stream
+          intptr_t   stream
 )
 {
     flow_rls_check_bending(bending, "flow_diag_rls");
@@ -402,7 +402,7 @@ void flow_relax_rls(
           int8_t     bound     ,
           int        ndim      ,
           int        nb_iter   ,
-          int        stream
+          intptr_t   stream
 )
 {
     flow_rls_check_bending(bending, "flow_relax_rls");
